requirements.txt
Flask==1.0.2
sympy==1.4
mysql_connector_python==8.0.18
